server/requirements.txt
Django==3.0.1
django-cors-headers==3.2.0
djangorestframework==3.11.0
djangorestframework-jsonapi==3.0.0
